MAINT, API: move _bit_generator to bit_generator
