+++ numpy/random/bit_generator.pxd
